Update atlas.min.js
This isn't really the minified version. Updated the behavior of the and/or collaboration search to a new button in the filters panel, cleaned up some of the behavior of filters, and added support for static table headers by making some minor changes to JQuery. Added comments to some functions.